Create music_accuracy.py
This program measures the accuracy of the model used to predict the genre of music liked by people depending on their gender and age (in music_prediction.py). It splits the data into two sets: one for training and one for testing, with 80% of the data used for training and 20% for testing. Instead of passing only two samples to make a prediction (as was done in music_predictor.py), it passes the dataset for testing, gets the predictions and the compares it with the actual values in the test set. Then, the accuracy is calculated and displayed.